+++ src/test/java/com/github/alexdlaird/ngrok/installer/NgrokInstallerTest.java
@@ -86,11 +86,11 @@ class NgrokInstallerTest extends NgrokTestCase {
 
         // WHEN
         ngrokInstaller.installNgrok(javaNgrokConfig.getNgrokPath(), javaNgrokConfig.getNgrokVersion());
-        ngrokProcessV2 = new NgrokProcess(javaNgrokConfig, ngrokInstaller);
+        ngrokProcessV3 = new NgrokProcess(javaNgrokConfig, ngrokInstaller);
 
         // THEN
         assertTrue(Files.exists(javaNgrokConfig.getNgrokPath()));
-        assertTrue(ngrokProcessV2.getVersion().startsWith("2"));
+        assertTrue(ngrokProcessV3.getVersion().startsWith("3"));
     }
 
     @Test